Completely refactored to simplify the flow and usage!!!! It now has a "inbox.google.com" feel
